requirements.txt
discord.py
requests
validators
